+++ CHANGELOG.md
@@ -4,7 +4,10 @@ Changelog
 Work in Progress
 ----------------
 
-  * Bugfixes.
+  * Bugfixes, expanding defaults.
+  * Change name of socket action to connect.
+  * Add connect and wait for response action.
+  * Expanded documentation.
 
 v0.0.1 - 04/08/2013
 -------------------